requirements.txt
discord.py
asyncpg
python-dotenv
aiohttp